requirements.txt
numpy>=1.19.2
streamlit>=1.22.0
opencv-python>=4.5.3
pandas>=1.3.0
tensorflow>=2.8.0
spotipy>=2.22.1
pytube>=12.1.0
python-dotenv>=0.19.0
requests>=2.27.1